requirements.txt
python==3.11.3
scikit-learn==1.2.2
pandas==2.0.2
matplotlib==3.7.0
requests==2.28.1
pyyaml==6.0
Flask==2.3.2
SQLAlchemy==2.0.15